Bug 1470587: Uninitialize reporter created during test. r=mossop
At some point, the matching call to reporter.uninit got removed from this test.
The result is that the reporter still exists and is still collecting errors
during the rest of the tasks in the file. In most tasks, this isn't an issue
since we use reporter.handleMessage to test message handling at a per-reporter
level.

But the telemetry measures are shared between multiple reporters, thus they are
susceptible to interference from other running reporter instances.

The error that is being logged when this test fails is from the test add-on
created in testAddonIDMangle. My best guess is that the error logged by the
add-on is being processed during an idle moment in another task, since we
schedule processing to be delayed until the browser is idle. It seems this
scheduling is pretty consistent on certain Linux platforms.

Differential Revision: https://phabricator.services.mozilla.com/D1836

UltraBlame original commit: d26f85971b0a14884ea88114b89e0622ab317cee