Bug 1340498 - Update onVisits uses to 'page-visited' r=mak
Consuming the new 'page-visited' notification was fairly trivial,
since it was already brought over to onVisits. There's not much to
say about this other than that I'm a little bit uncertain about
all the hoops we have to jump through to get a JSContext and
GlobalObject from History.cpp (which is discussed in the earlier
commit in the series).

MozReview-Commit-ID: LHaBWSylyLI

UltraBlame original commit: b6f047709e8e11202f5e81dee8cc26620595f4d1