Bug 1459900 - Track structure field mutability. r=luke
The StructType and AstStructType machinery must retain the information about
field mutability (to be used by subsequent patches).

For StructType, I opted to create a StructField structure that holds the
information per field.  This works well except when we (occasionally, not in
this patch) need an array of ValTypes for some existing functionality.

For AstStructType, I opted to keep the array of mutability flags separate from
the arrays of names and field types, because this was simplest.

In either case, we could do it the other way.

UltraBlame original commit: 3ed7ed0f031b842901bee0dbc27c7a3a39ed16ae