Bug 1471941 - Clean up dangling stylesheet listener in DevTools. r=ochameau
In bug 1465637, we removed the redundant actor pool that was used only for
creating stylesheet actors in a target.  This change happened to lead to errors
that uncovered a dangling listener in `StyleSheetsActor` that would continue to
create actors after toolbox destroy when a new sheet is added.

Differential Revision: https://phabricator.services.mozilla.com/D1866

UltraBlame original commit: 30bdb22bbe5a2bdecd09a188b78850bd460b4fae