Bug 1474258 - Remove the "listheader" binding and move the remaining styles to "richlistbox.css". r=bgrins
The content of this element is always overridden and the "xul:button" display mode is unnecessary, so the binding can be removed. The styles that control color are moved to "richlistbox.css" in preparation for the removal of the "listbox" element.

MozReview-Commit-ID: 5pGVE6n34EL

UltraBlame original commit: bd281782dc93474944f6a9375312e7e3a31cd061