Bug 1474814: Invalidate the AccessibleHandler cache on focus events. r=MarcoZ
We don't fire a state change for the focused state, since that would be redundant given that we fire a focus event.
However, that also means the handler cache previously didn't get invalidated when there was a focus change without any other update.
That resulted in stale focused states, which could mean an object was exposed as focused when it wasn't.
Invalidating the cache for focus events ensures the focused state is always up to date.

MozReview-Commit-ID: DZTJMbiTwgt

UltraBlame original commit: cbf89941ccff2dde578bb11be34aec1a1b902b1f