Bug 1423011 - Part 2: Add gtests. r=botond
Ensures that APZ correctly updates the layout viewport offset when the visual
viewport exceeds the boundaries of the layout viewport.

Includes changes to APZCPinchTester::GetPinchableFrameMetrics to correct the
offset of the composition bounds (which should always be (0, 0) for the
RCD-RSF) and fixes to impacted test cases.

MozReview-Commit-ID: I3hTx9kCjEP

UltraBlame original commit: 639b78b577608511cbe98ccb151b0a043d23a564