Bug 1475165 - Fix inspector tests; r=gl.
Adapt inspector test to the new order of the search results.
This was an easy task since we were already flipping the
expected results to make the test more readable (which makes
me think that this is the "natural" order).

MozReview-Commit-ID: DpxW2O1eOi4

UltraBlame original commit: 69e53775fd5cabcab8c87c72466eb39ab542b4bf