Bug 1354232 - Allow wpt manifest files to specify LSAN errors to ignore, r=maja_zf
This adds a property lsan-allowed to the expectation manifest files
that takes a list of strings. Any entry in the list that matches a
frame in an LSAN stack will cause that stack to be regarded as an
expected failure.

MozReview-Commit-ID: 2oUw0joThha

UltraBlame original commit: b8da0ef2695d8bca1c11b44447a71cef5ce5b3fc