Bug 1354232 - Add support for updating LSAN data in wpt-update r=maja_zf
LSAN data differs from existing expectation data because the data is
only generated when the browser exits, so the problems reported can
happen at any point in the current session. We use the `scope`
property in the log message to determine the path to a __dir__.ini
file that covers all the tests run in the session, and add the LSAN
exclusion rules to there.

The rules themselves are generated by taking the topmost frame of any
stack that's reported as unexpectedly leaking, and adding that to the
list of permitted frames in the lsan-allowed property. We never remove
entries from this list since intermittents may be present which won't
appear on a specific run. Instead we rely on humans fixing the issues
to also clean up the expectation files.

MozReview-Commit-ID: Kxm0hFXlGE3

UltraBlame original commit: 9e8a66385f66f89f8cf2c9a18edf008496fa89fa