Bug 1476238 - Use nsresult/promise rejection in initListBuild r=Gijs
I mistranslated the boolean success value of the old initListBuild
method. Simplifying it to just use a promise rejection should be
equivalent and simpler, since this is only run from the update
method of WindowsJumpLists, and the call sites for that don't do
anything afterwards.

Differential Revision: https://phabricator.services.mozilla.com/D2244

UltraBlame original commit: 2727ea579453784ab7933bf2ca4390d180211651