Bug 1477930 - Remove taggingService HasTags synchronous API. r=Standard8
The tagging API is moving to the bookmarking API, this is part of it.
This will break the tags root open glyph in the Library, for now we don't have
an alternative, we'll re-evaluate in the future in bug 1477934.

Differential Revision: https://phabricator.services.mozilla.com/D2318

UltraBlame original commit: 285d7793a541ee131f12a761ac96d5c194ea7b05