Bug 1476263 - Suggest to create mozconfig if it does not exist. r=gps
A frequent question when mentoring new contributors is what "your
mozconfig file" is.  By suggesting to create the file if it does
not exist, we can hopefully alleviate some new contributor frustration.

This change does unfortunately not take into account that the
mozconfig file can be named .mozconfig or even be in a designated
location defined by the MOZCONFIG environment variable, but it
seems reasonable to assume that developers who already know about
those alternatives will know which file to edit, and that what we
should optimise for during the bootstrapping process is to get new
contributors up and running quickly.

UltraBlame original commit: 953910565893b2da0891a9d033d3fa0a6bad8b81