Bug 1477041 - Modified MediaControlService and notification flow. r=jchen
Created a new parcelable which is sent to MediaControlService where
the notification is created based on the data of this object instead of
referencing a static notification. Toggling the service is now done through
startService with a specific action for shutting down instead of calling
stopService due to concurrency issues where the service can be stopped before
having a chance to call startForeground.

MozReview-Commit-ID: 6qNPintkVy

UltraBlame original commit: 06cdeb2fd5224952178ff7a804b9151a1b5f2af9