Bug 1479681: Fix the loop in nsTransitionManager dealing with stopping `all` transitions. r=hiro a=aryx
The loop was mutating the nsCSSPropertyID used to guard the exit, which is
obviously wrong.

This branch is pretty rarely taken, since people don't usually specify `all` as
a transition property other than the first, for which case we take the fast path
with `checkProperties = false`. Our test-suite failed to catch this.

Added a crashtest that hangs without this patch.

The reason bug 1478990 regressed this is because it changed the order of
nsCSSPropertyID so that `p` actually went backwards causing the infinite loop,
but the bug was introduced (by me, whoops) in bug 1309752.

Differential Revision: https://phabricator.services.mozilla.com/D2552

MozReview-Commit-ID: Ii3D1FaZ31R

UltraBlame original commit: 0d72c7996d60a7c07e35c5f90d78b02a47d17460