Bug 1478305 - Update ESLint to treat ChromeUtils.import definitions as real variable definitions for single-export modules. r=mossop
One minor issue with this method, is that we are unable to support exported definitions without further work. This may cause false-positives if items are exported. However, this is already the case with the other import mechanisms, and the benefits seem to be larger than the disadvantages here.

MozReview-Commit-ID: 4fx2aLRBt7T

UltraBlame original commit: 0fd5b04d2dee99a4d060998dab164294e09ceb9d